requirements.txt
numpy==1.20.3
Pillow==9.3.0
scipy==1.8.1
torch==1.8.0+cu111
torchvision==0.9.0+cu111
tqdm==4.62.3
